+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 # Changelog
 
 ## vNEXT
-
+- Add `dealId` to `TaskClaimedWithVoucher` event. (#61)
 - Add type-checking script (#53)
 - Run partial upgrade tests on fork.
-  - VoucherHub
-    - [Should upgrade decimals](test/VoucherHubUpgrade.test.ts) (#58)
+    - Should maintain consistent voucher addresses (#61)
+    - Should upgrade decimals (#58)
 - Implement `VoucherHub` upgrade and test it on GitHub Action CI using a local fork. (#57)
 - Relocate `decimals()` function in VoucherHub smart contract. (#56)
 - Better handling for env variables. (#55)